Separate main command and cli command
