+++ cmd/ogen-cli/rpcclient/network.go
