+++ cmd/ogen-cli/rpcclient/wallet.go
