+++ cmd/ogen/cli/generatekeys.go
