+++ cmd/ogen/cli/reset.go
