+++ scripts/build.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 
 go build cmd/ogen/ogen.go
+go build cmd/ogen-cli/ogen-cli.go
 go build cmd/migration/migration.go
